requirements.txt
playwright
fastapi
uvicorn
python-dotenv
selectolax
pydantic
pandas
openpyxl
tenacity
requests
sqlalchemy
PyYAML
python-multipart
cloudinary
